AnsibleModule - add methods for backwards compatibility
With the move the the ArgumentSpecValidator class in AnsibleModule, many private methods
were removed since that functionality moved to the validator class. However, this breaks
existing plugins that are using the private methods.

While the methods are private and leaving dead code around is undesireable, leaving the private
methods around does help ease the transition to the new interface.