task_queue_manager: Wait for clean termination of worker processes via join()
Do not unconditionally terminate() running worker processes because they
may still hold the internal locks of the _final_q.

The documentation [1] warns about using terminate() on processes that may
hold locks or semaphores:

    Warning: If this method is used when the associated process is using a
    pipe or queue then the pipe or queue is liable to become corrupted and
    may become unusable by other process. Similarly, if the process has
    acquired a lock or semaphore etc. then terminating it is liable to cause
    other processes to deadlock.

The multiprocessing.Queue() uses a Semaphore lock inside. So if the worker
was killed while it was in the process of running put(), the next worker
would see the internal semaphore locked and unable to use the _final_q anymore.

This uses 60 seconds as a timeout. This should be plenty, _cleanup_processes()
is only reached when all results from the workers have been received through
the queue and the workers are about to exit.

[1] https://docs.python.org/3/library/multiprocessing.html#multiprocessing.Process.terminate