Lazy-load Templar when pending results
In the default configuration (internal poll interval set to 0.001),
Ansible consumes a significant amount of CPU while waiting for results.
On my machine [1], about 23-25% of one CPU is used when waiting.

Ansible spends a lot of resources on the instanciation of Templar, while
in most iterations it is not useful since there is no result to return.

By lazy-loading the `handler_templar`, the CPU usage decreases to
6-8% with the default configuration.

[1] i5-8365U CPU @ 1.60GHz