Fix the Makefile's htmlsingle command.
Rather than documenting that the end user needs to run two make targets
to test changes to a file, modify the Makefile so that htmlsingle
invokes the other make target.

Change the dependencies for the *_structure targets.  They don't depend
on generating the module and plugin docs so move the dependency to a
higher level target instead.  This way, the htmlsingle target doesn't
need to run those.

Add make to the list of dependent packages needed to build the docs.