version test - improve message when value is empty
When an empty value is provided, no `version` attribute will exist on the `LooseVersion` or
`StrictVersion` object. We catch and handle this, but it's not immediatebly clear that an
AttributeError means an empty value was provided.

Handle this case specifically and add a more helpful error message.

Add unit and integration tests.