first_found: clear up the skip option usage
I don't think I'm the first person to be confused by the "skip"
semantics here; the option was deprecated with
e17a2b502d6601be53c60d7ba1c627df419460c9 but then restored with
707e33793d683254a511cd1ec825df86a5121feb.

This confusion seemed to stem from the two two ways this can be used;
when used with "with_first_found", setting "skip: True" lives up to
its name and makes the task skip when no files are found.

When used in a template with lookup()/query(), however, "skip: True"
returns an empty list.  You have plenty of options...

  lookup('first_found', dict(files=['notfound']), errors='ignore') -> ""
  lookup('first_found', dict(files=['notfound'], skip=True))       -> []
  lookup('first_found', dict(files=['notfound']), wantlist=True)   -> []
  q('first_found', dict(files=['notfound']), errors='ignore')      -> []

I think this is not helped by the extant error that mixes things up
saying "use errors='ignore' to skip" which, as noted, doesn't actually
skip.

This clarifies the documentation for these range of cases.  It adds
more examples and reduces the error to just "no file was found" to
avoid giving wrong suggestions.