Use a manager object when storing module locks
This change will allow Ansible to use a manager object when
multiprocessing to ensure cached modules are cached as expected,
and are not creating multiple locks when multiprocessing, which
can result the race condition  `FileExistsError: [Errno 17] File exists`.

Module cache will now be stored in the manager, using the same fork
multiprocessing context which will now allow `Locks` to be shared
across processes. Because locks are only intended to be shared through
inheritance, all locks are also using a manager and built using the
same global manager as the module container.

- Tests have been added to ensure the manager dictionary is
  functioning as expected.

- This change adds a couple action constants which were missing from
  the constants file.

- This change only resolves the locking problems when running under py3.
  In py2 the original code remains effective, meaning cached modules will
  still not be a member of the write lock container when added from within
  a sub-process.

- Bugfix Pull Request

rhbz: 1950050
Co-authored-by: Sviatoslav Sydorenko <wk.cvs.github@sydorenko.org.ua>
Signed-off-by: Kevin Carter <kecarter@redhat.com>