Modify rpm_key to check for multiple fingerprints in keyfiles
Multiple keys can be concatenated into a single file however. This aims
to make the module check for multiple key fingerprints beng present in a
single file. This also allows for a list of fingerprints to be passed to
module arguments for verification. If there's any mismatch, then the
module throws an error.