regex_search: always return an empty string when no match is found
The docs state that regex_search returns an empty string when no match
is found. This is a side-effect of our custom implementation of
_finalize method for Jinja's Environment where we explictly convert any
result of templating that is None to an empty string. Which is a case of
regex_search function that the filter uses because it returns None when
no match is found.

The problem here is when the filter is used inline with any comparison
to an empty string like the following:
``{{ 'hello' | regex_search('world') == '' }}``

This would be evaluated to False because the _finalize method is only
called on the final result of templating as opposed to on the
itermediate results, so it would effectivelly be ``None == ''``.

This patch fixes that to explictly return an empty string when no match
is found in the regex_search function.