Document return values of template module
Add documentation for some return values of template module. I've only
included values that are returned consistently, there are more that are
only returned under special circumstances - like when destination is
changed - but I'm not entirely sure if that behavior is entirely
intentional or subject to change.

I have not tested this extensively, so it's possible that under special
circumstances this might not be entirely accurate, but briefly looking
at the code as far as I can tell they are correct.

Sample code I used to see the returned values:

```
$ ansible -m template -a "dest=/tmp/test src=/tmp/template.j2" localhost
localhost | SUCCESS => {
    "changed": false,
    "checksum": "373296322247ab85d26d5d1257772757e7afd172",
    "dest": "/tmp/test",
    "gid": 1000,
    "group": "mprasil",
    "mode": "0664",
    "owner": "mprasil",
    "path": "/tmp/test",
    "size": 22,
    "state": "file",
    "uid": 1000
}

```