honor `retries` even if `until` is absent
this is one step to allowing a bare `retries` to default
to using `until: {{ task }} is succeeded`.
We need to be able to distinguish whether `retries` has been manually specified.
This requires 2 changes:
  - removing the default for the variable
  - assuming that `retries: null` is equivalent to not specifying a value for `retries` (or vice versa)

So I have removed the default for that variable.
Removing the default value is acceptable to me.
It's use here is overridden and it isn't used elsewhere.

One consequence of treating `retries: null` as unspecified
is that will now result in 4 attempts, and not 3.
(It is now not possible to distinguish those cases)
I think that's fine, as `retries: null` is either nonsensical
or an explicit request for default.

The alternative is to use a sentry value for the default,
but that seems the _most_ risky option,
since the type will not match the expected type
but will not evaluate as None