modules apt: Force update if default_release is not available
This patch fixes apt_pkg.Error which is triggered on cache.open() when
three conditions are satisfied:

* An repository with a new release name (such as 'buster-backports') is
  added.
* The apt cache is not up-to-date (i.e. no
  /var/lib/apt/lists/*buster-backports* file is found).
* Ansible apt module is invoked with default_default set to this new
  release.

Even if update_cache is set to yes the apt cache is not update and the
module fails with this apt_pkg.Error error:

  E:The value 'buster-backports' is invalid for APT::Default-Release as
    such a release is not available in the sources

The solution is to force a cache update if default_release is set and
apt_pkg.Error is raised.

This bug happens with python3-apt 1.8.4.3 and maybe later versions. It
can be safely kept for all python3-apt versions since it make no sense
to install package if the related list file is not found in
/var/lib/apt/lists.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>