dnf: allow handling of modules without a default profile
DNF modules can be shipped without a default profile. In this case
calling 'install module:stream' errors out, as DNF doesn't know
what to install (as this is determined by the profile).

One solution is to explicitly call 'install module:stream/profile'.
However, it can be beneficial to only enable a module, without
installing anything from it. This is useful when the module is just
needed for dependencies, and maybe doesn't ship *any* profiles at all.

Examples for such modules are pki-deps and pki-core in EL8.

Fixes: #56504