Update intro_inventory.rst
I believe the first statement is clear enough, and that the second statement only creates confusion of where additional `group|host_vars` directories should be placed. As I read the second statement, I would expect the `group|host_vars` directories to be placed where the `ansible-playbook` command is issued. In practice, I have found that this is incorrect, and that the `group|host_vars` directories need to be adjance to the playbook that is being executed.

For example, executing `ansible-playbook playbooks/pb.yml` should have the `group|host_vars` directories inside the `playbooks` directories. Per the second statement, I would actually expect `group|host_vars` directories to be placed in the directory that the command is executed from.