Fix documentation around the password parameter of builtin.user
The documentation said that setting the password to '!' or '*' on Linux
or '*************' on OpenBSD would create a disabled account, which I
think is incorrect. So fix the documentation so that it says that using
those values creates an account with a locked/disabled password.

You can find detailed background information on the wording as well as
detailed examples in the accompanying PR.
However here are some places to look at for the difference between a
disabled account and an account with a locked/disabled password:
  - Linux - man passwd.1: The -l option.
  - Linux - man passwd.5: The paragraph beginning with "If the password
    field contains some string...".
  - Linux - man usermod.8 The -L, --lock option.
  - OpenBSD - man passwd.5: The sentence beginning with `Note that there
    is nothing special..." and the following sentence.