Added package reference for Debian to resolve error
##### SUMMARY
Added a reference that in Debian systems, this error can be quickly and easily resolved by installing the `acl` package, which allows / implements the `setfacl` command. This error comes up when using the community.postgres collection. That collection points to using become_user: postgres, but then the "Failed to set permissions on the temporary files Ansible needs to create when becoming an unprivileged user" error occurs, and points to this document to resolve that error.


##### ISSUE TYPE
- Docs Pull Request

+label: docsite_pr