copy module - fix reporting changed for src directories
Return 'state' result key consistently

Ensure 'dest' reflects the invocation when it's a directory

Check for a single file or symlink when using the module_return to update the action plugin result.
Before, if a file and symlink existed, the symlink module_return was used to update the result,
but if only a symlink existed, its module_return would not be used to update the result.
Now one file or one symlink are consistent.