collect printing of role info
Harmonise on simply using `role_path` for both showing single and listing all
instead of `os.path.dirname(GalaxyRole.path)` for showing and `role_path` for listing

`gr.path` is equivalent, but takes a bit of a longer path. When listing a single item,
we pass in `os.path.join(role_path, role_name)` to `GalaxyRole.__init__` as the path.
This path has some extra logic which is skipped
(the path is proper and ends with the role_name by construction).
We the remove the rolename from the path with `os.path.dirname`,
effectively giving us `role_path` back.