fix bug in _get_collection_widths
function expects a collection-like argument,
but will also wrap a single element into a tuple.
However, it checks for whether the argument is an iterable.
The expected type (`Candidate`) is a namedetuple, which is an iterable,
so the promotion would not recognise this.
There was a test for this case, but it monkeypatched the `is_iterable` call,
presumably because the `Candidate` or `Requirement` _are_ iterables
and this gets the test to pass.

Solution is to only wrap the classes we expect to receive as single elements.