callback: query strategy whether to cache taskname
With the free strategy, the last task name should not be cached. Ref: #43950

However, the same might be true of a strategy plugin, a popular example being
mitogen_free. Otherwise, task names will be incorrect for mitogen_free just
like in #43950

Rather than checking for the name of the strategy in the callback plugin, we
can query the strategy itself whether task names should be cached or not.

Since the strategy is not available directly from the callback, passing a flag
on the play object via v2_playbook_on_play_start. A play has only one strategy.