Action plugin pre-encoding hook
`_configure_module` is a critical but very monolithic method. A new ActionModule doing interesting stuff, as elegant be it, is likely to copy/paste much of its code.  
This method is also responsible of calling `modify_module` after which module's arguments are definitely encoded.

An ActionModule could legitimately be interested into accessing both `become` and `modules_*` variables at that point (just before it's too late).

The present change moves the small code block about `become_kwargs` inside a dedicated method, thus allowing an ActionModule to override it.
This newly introduced method takes as arguments multiple variables not currently part of `self`. The default behavior is totally preserved but such a new method gives a chance to an ActionModule to modify them.

This trivial change greatly opens up extensibility of ActionModule in the most needed code-path.
Bonus point: It may make that part of the code easier to unit-test.

Warning: draft/WIP RFC-style PR.