Point to Kerberos as a safer winrm setup method
##### SUMMARY
<!--- Your description here -->
Only noting that the provided 'ConfigureRemotingForAnsible.ps1' script is not safe for production use, is not a very good default. The majority of installs will use the provided 'ConfigureRemotingForAnsible.ps1' script and at best only use basic auth, which is very bad from a security stand point and is also being depreciated left right and center. We should really point to where we describe how to do this more securely for live production installs.


##### ISSUE TYPE
- Docs Pull Request

+label: docsite_pr