Replace `mock` with `unittest.mock`.
mock is a backport of the standard library unittest.mock module, which
is available in Python 3.3 and above. I do not think there is a
compelling reason to use mock when we only support Python 3.8 and above.
Prior to the removal of units.compat.mock
(https://github.com/ansible/ansible/pull/77118/), `mock` was only used
when `unittest.mock` was unavailable, but now `mock` is unconditionally
used.

Furthermore, `mock` has been deprecated in Fedora and was completely
removed from RHEL 9, as it's redundant for the versions of python
available in those distributions. Alongside the Fedora specific
information, there is some general information about this issue in the
[Deprecate python-mock Change
Proposal](https://fedoraproject.org/wiki/Changes/DeprecatePythonMock)
which you can take a look at.

Signed-off-by: Maxwell G <gotmax@e.email>