Update vault.rst
I read this documentation and felt that it was unclear how to provide a vault file password by any means other than a command line option.  The information is there but it is a bit obscure.  I propose this change to fix that.

I also note that the paragraph following "Using encrypted variables and files" says you can set the password file in a playbook - I don't think that information on how to do this is on this page.  If it is I couldn't find it.